LayoutTests/platform/mac/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-audio-element/audio-loading-load-preload-metadata-deferred.tentative-expected.txt



FAIL Audio with loading=lazy and preload=metadata that is not visible in viewport does not load audio data assert_equals: expected 0 but got 1
